Add geolocator functions
Your package seems to duplicate much of the functionality of our censusr package (on CRAN), but we think yours is getting a good foot hold and has many features that we like more. So we'll be deprecating our package. 

In the meantime, here are a couple of functions to geolocate addresses or lat/long points that use the Census API. Unless you think those are more in line with `tigris`?